chore: rename entities and add to config
