+++ api/src/config/ormconfig.ts
@@ -4,6 +4,8 @@ import { MediaEntity } from "../entity/media.entity";
 import { CommentEntity } from "../entity/comment.entity";
 import env from "../../environment";
 import { NotificationEntity } from "../entity/notification.entity";
+import { WatchListEntity } from "../entity/watchList.entity";
+import { WatchListItemEntity } from "../entity/watchListItem.entity";
 
 const appDataSource = new DataSource({
     type: "postgres",
@@ -12,7 +14,7 @@ const appDataSource = new DataSource({
     username: env.DB_USER,
     password: env.DB_PASSWORD,
     database: env.DB_NAME,
-    entities: [UserEntity, MediaEntity, CommentEntity, NotificationEntity],
+    entities: [UserEntity, MediaEntity, CommentEntity, NotificationEntity, WatchListEntity, WatchListItemEntity],
     migrations: ["src/migration/**/*.ts"],
     synchronize: true,
 });
